feat: colocate with routes
