--- apps/sage/src/constants/routes.ts
@@ -1,5 +1,0 @@
-export const routes = {
-  home: "/",
-  agentDetail: "/agents/$agentId",
-  agentRuns: "/agents/$agentId/runs/$runId",
-} as const;
