--- apps/sage/src/pages/Home/index.tsx
@@ -1,8 +1,0 @@
-import { TestComponent } from "components/TestComponent";
-
-export const HomePage: React.FC = () => (
-  <div>
-    <p>Home page</p>
-    <TestComponent />
-  </div>
-);
