+++ apps/sage/src/routes/agents/$agentId/index.tsx
@@ -1,5 +1,11 @@
-import { createFileRoute } from "@tanstack/react-router";
-import { AgentDetailPage } from "pages/AgentDetail";
+import { createFileRoute, getRouteApi } from "@tanstack/react-router";
+
+const routeApi = getRouteApi("/agents/$agentId/");
+
+const AgentDetailPage: React.FC = () => {
+  const { agentId } = routeApi.useParams();
+  return <p>Agent ID: {agentId}</p>;
+};
 
 export const Route = createFileRoute("/agents/$agentId/")({
   component: AgentDetailPage,
