+++ apps/sage/src/routes/agents/$agentId/runs/$runId.tsx
@@ -1,5 +1,15 @@
-import { createFileRoute } from "@tanstack/react-router";
-import { AgentRunsPage } from "pages/AgentRuns";
+import { createFileRoute, getRouteApi } from "@tanstack/react-router";
+
+const routeApi = getRouteApi("/agents/$agentId/runs/$runId");
+
+const AgentRunsPage: React.FC = () => {
+  const { agentId, runId } = routeApi.useParams();
+  return (
+    <p>
+      Agent ID: {agentId}, Run ID: {runId}
+    </p>
+  );
+};
 
 export const Route = createFileRoute("/agents/$agentId/runs/$runId")({
   component: AgentRunsPage,
