+++ apps/sage/src/routes/index.tsx
@@ -1,5 +1,12 @@
 import { createFileRoute } from "@tanstack/react-router";
-import { HomePage } from "pages/Home";
+import { TestComponent } from "components/TestComponent";
+
+const HomePage: React.FC = () => (
+  <div>
+    <p>Home page</p>
+    <TestComponent />
+  </div>
+);
 
 export const Route = createFileRoute("/")({
   component: HomePage,
